requirements.txt
asgiref==3.5.0
defusedxml==0.7.1
diff-match-patch==20200713
django>=3.2,<3.3
django-bootstrap3==21.2
django-crispy-forms==1.14.0
django-filter==21.1
django-import-export==2.7.1
django-tables2==2.4.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.0
xlrd==2.0.1
xlwt==1.3.0
gunicorn
psycopg2-binary